requirements.txt
cycler==0.11.0
decorator==4.4.2
freetype-py==2.2.0
imageio==2.15.0
kiwisolver==1.3.1
matplotlib==3.3.4
networkx==2.5.1
numpy==1.19.5
opencv-python==4.5.5.64
Pillow==8.4.0
plotext==4.2.0
pyglet==1.5.23
PyOpenGL==3.1.0
pyparsing==3.0.8
pyrender==0.1.45
python-dateutil==2.8.2
scipy==1.5.4
six==1.16.0
trimesh==3.10.8
